+++ src/saltchannel/util/Bytes.java
@@ -1,6 +1,6 @@
 // Copyright Frans Lundberg, Stockholm 2016.
 // Public domain code copied from https://github.com/franslundberg/java-cut/ 
-// October 2016.
+// October 2016. Tested there.
 
 package saltchannel.util;